git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@8571 01821946-851b-0410-b7f4-f1477c70e57a

+++ network/multi.cpp
@@ -2364,7 +2364,7 @@ bufI += 2;
 PUT_INTEL_INT (gameData.multigame.msg.buf + bufI, LOCALPLAYER.flags);
 bufI += 4;
 #if 1
-if (multiMessageLengths [nType][1] > 0) {
+if (multiMessageLengths [(int) nType][1] > 0) {
 	if (gameStates.multi.nGameType == UDP_GAME) {
 		PUT_INTEL_SHORT (gameData.multigame.msg.buf + bufI, gameStates.app.nRandSeed);
 		bufI += 2;
